fix 103

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,17 +8,17 @@
 /* betty style doc for function main goes there */
 int main(void)
 {
-	double i = 1, j = 1, temp, sum = 0;
+	unsigned long int i = 1, j = 1, temp, sum = 0;
 
 	while (j < 4000000)
 	{
-		if (j % 2.0 == 0)
+		if (j % 2 == 0)
 			sum += j;
 		temp = i;
 		i = j;
 		j = temp + i;
 	}
-	printf("%.0f\n", sum);
+	printf("%d\n", sum);
 
 
 	return (0);
